+++ cypress/e2e/user-journeys/custodians/users.cy.ts
@@ -20,14 +20,21 @@ import { DEFAULT_PROJECT_INVITE_USERS } from "cypress/support/utils/data";
 const dataProjectInviteUser = DEFAULT_PROJECT_INVITE_USERS;
 
 describe("Projects custodians journey", () => {
-  beforeEach(() => {
+  it('should have no detectable accessibility violations on load', () => {
+      loginCustodian();
+      cy.visitFirst(ROUTES.profileCustodianUsers.path);
+      cy.waitForLoadingToFinish();
+      cy.checkA11yPage();
+    });
+})
+
+
+describe("Projects custodians journey", () => {
+ beforeEach(() => {
     loginCustodian();
-
     cy.visitFirst(ROUTES.profileCustodianUsers.path);
-
     cy.contains("a", getName(dataProjectInviteUser)).click();
   });
-
   before(() => {
     loginCustodian();
     goToProjectUsersList();
@@ -37,7 +44,7 @@ describe("Projects custodians journey", () => {
   after(() => {
     logout();
   });
-
+  
   it("Has Affiliations Tab Custodian User", () => {
     hasAffiliationsTabCustodianUser();
   });