+++ src/flags.ts
@@ -1,0 +1,19 @@
+import { flag } from "@vercel/flags/next";
+import { createAPIFlagAdapter } from "./utils/apiFlagAdapter";
+
+const apiAdapter = createAPIFlagAdapter();
+
+export const isTestFeatureEnabled = flag({
+  key: "test-feature",
+  adapter: await apiAdapter(),
+});
+
+export const isChristmasBannerEnabled = flag({
+  key: "christmas-banner",
+  adapter: await apiAdapter(),
+});
+
+export const isTestFeatureUserAdmin = flag({
+  key: "test-feature-user-admin",
+  adapter: await apiAdapter(),
+});