+++ src/modules/BannerMessage/BannerMessage.tsx
@@ -1,10 +1,20 @@
-import { mockedBannerContent } from "@/mocks/data/cms";
-import { Message, MessageProps } from "../../components/Message";
+import { mockedChristmasBannerContent } from "@/mocks/data/cms";
+import { BannerLists, Message, MessageProps } from "../../components/Message";
 
-export default function BannerMessage(props: MessageProps) {
+export interface BannerMessageProps extends MessageProps {
+  enabledBanners: BannerLists;
+}
+export default function BannerMessage(props: BannerMessageProps) {
+  const { enabledBanners, ...messageProps } = props;
+  const { christmasMessage } = enabledBanners;
+
   return (
-    <Message variant="filled" severity="warning" {...props}>
-      {mockedBannerContent()}
+    <Message
+      variant="filled"
+      severity="warning"
+      data-testid="banner"
+      {...messageProps}>
+      {christmasMessage && mockedChristmasBannerContent()}
     </Message>
   );
 }